# Remove nohash-hasher dependency