--- a/Cargo.toml
+++ b/Cargo.toml
@@ -251,7 +251,6 @@ lightningcss = { version = "1.0.0-alpha.57", features = [
   "into_owned",
 ] }
 mime = "0.3.16"
-nohash-hasher = "0.2.0"
 notify = "6.1.1"
 notify-debouncer-full = "0.3.1"
 notify-debouncer-mini = { version = "0.3.0", default-features = false }