--- a/crates/turbo-tasks-memory/src/count_hash_set.rs
+++ b/crates/turbo-tasks-memory/src/count_hash_set.rs
@@ -363,13 +363,11 @@ impl<'a, T> Iterator for CountHashSetIter<'a, T> {
 
 #[cfg(test)]
 mod tests {
-    use nohash_hasher::BuildNoHashHasher;
-
     use super::*;
 
     #[test]
     fn test_add_remove() {
-        let mut set: CountHashSet<i32, BuildNoHashHasher<i32>> = CountHashSet::new();
+        let mut set: CountHashSet<i32> = CountHashSet::new();
         assert_eq!(set.len(), 0);
         assert!(set.is_empty());
 
@@ -413,7 +411,7 @@ mod tests {
 
     #[test]
     fn test_add_remove_cloneable() {
-        let mut set: CountHashSet<i32, BuildNoHashHasher<i32>> = CountHashSet::new();
+        let mut set: CountHashSet<i32> = CountHashSet::new();
         assert_eq!(set.len(), 0);
         assert!(set.is_empty());
 
@@ -457,7 +455,7 @@ mod tests {
 
     #[test]
     fn test_add_remove_if_entry() {
-        let mut set: CountHashSet<i32, BuildNoHashHasher<i32>> = CountHashSet::new();
+        let mut set: CountHashSet<i32> = CountHashSet::new();
 
         assert!(!set.add_if_entry(&1));
         assert_eq!(set.len(), 0);
@@ -487,7 +485,7 @@ mod tests {
 
     #[test]
     fn test_zero() {
-        let mut set: CountHashSet<i32, BuildNoHashHasher<i32>> = CountHashSet::new();
+        let mut set: CountHashSet<i32> = CountHashSet::new();
 
         assert!(!set.add_count(1, 0));
         assert_eq!(set.len(), 0);