--- a/crates/turbo-tasks/Cargo.toml
+++ b/crates/turbo-tasks/Cargo.toml
@@ -34,6 +34,7 @@ once_cell = { workspace = true }
 parking_lot = { workspace = true }
 pin-project-lite = { workspace = true }
 regex = { workspace = true }
+rustc-hash = { workspace = true }
 serde = { workspace = true, features = ["rc", "derive"] }
 serde_json = { workspace = true }
 serde_regex = "1.1.0"