--- a/crates/turbo-tasks/Cargo.toml
+++ b/crates/turbo-tasks/Cargo.toml
@@ -29,7 +29,6 @@ event-listener = "2.5.3"
 futures = { workspace = true }
 indexmap = { workspace = true, features = ["serde"] }
 mopa = "0.2.0"
-nohash-hasher = "0.2.0"
 once_cell = { workspace = true }
 parking_lot = { workspace = true }
 pin-project-lite = { workspace = true }