--- a/crates/turbo-tasks/src/lib.rs
+++ b/crates/turbo-tasks/src/lib.rs
@@ -73,6 +73,8 @@ mod value;
 mod value_type;
 mod vc;
 
+use std::hash::BuildHasherDefault;
+
 pub use anyhow::{Error, Result};
 use auto_hash_map::AutoSet;
 pub use collectibles::CollectiblesSource;
@@ -91,9 +93,9 @@ pub use manager::{
     TurboTasksCallApi, Unused, UpdateInfo,
 };
 pub use native_function::NativeFunction;
-use nohash_hasher::BuildNoHashHasher;
 pub use raw_vc::{CellId, RawVc, ReadRawVcFuture, ResolveTypeError};
 pub use read_ref::ReadRef;
+use rustc_hash::FxHasher;
 pub use state::State;
 pub use task::{
     concrete_task_input::{ConcreteTaskInput, SharedReference, SharedValue},
@@ -110,7 +112,7 @@ pub use vc::{
 
 pub use crate::rcstr::RcStr;
 
-pub type TaskIdSet = AutoSet<TaskId, BuildNoHashHasher<TaskId>, 2>;
+pub type TaskIdSet = AutoSet<TaskId, BuildHasherDefault<FxHasher>, 2>;
 
 pub mod test_helpers {
     pub use super::manager::{current_task_for_testing, with_turbo_tasks_for_testing};