--- a/crates/turbo-tasks/src/manager.rs
+++ b/crates/turbo-tasks/src/manager.rs
@@ -2,7 +2,7 @@ use std::{
     borrow::Cow,
     cell::RefCell,
     future::Future,
-    hash::Hash,
+    hash::{BuildHasherDefault, Hash},
     mem::take,
     panic::AssertUnwindSafe,
     pin::Pin,
@@ -17,7 +17,7 @@ use std::{
 use anyhow::{anyhow, Result};
 use auto_hash_map::AutoMap;
 use futures::FutureExt;
-use nohash_hasher::BuildNoHashHasher;
+use rustc_hash::FxHasher;
 use serde::{de::Visitor, Deserialize, Serialize};
 use tokio::{runtime::Handle, select, task_local};
 use tracing::{info_span, instrument, trace_span, Instrument, Level};
@@ -266,7 +266,7 @@ task_local! {
     /// The current TurboTasks instance
     static TURBO_TASKS: Arc<dyn TurboTasksApi>;
 
-    static CELL_COUNTERS: RefCell<AutoMap<ValueTypeId, u32, BuildNoHashHasher<ValueTypeId>, 8>>;
+    static CELL_COUNTERS: RefCell<AutoMap<ValueTypeId, u32, BuildHasherDefault<FxHasher>, 8>>;
 
     static CURRENT_TASK_ID: TaskId;